requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
pandas>=2.1.4,<3.0.0
requests>=2.31.0,<3.0.0
python-multipart>=0.0.6
pydantic>=2.5.2,<3.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.1,<4.0.0 